refactor(runtime): relocate v2 search DTOs to proximadb-runtime (TD-FLIGHT-1 step 2)
Move RichSearchRequest/Response/Result and RichFilterCondition/Operator from
the root-internal services::operations::vectors::legacy module into the platform
proximadb-runtime crate (new rich_search module), mirroring the TD-104 rich_record
relocation. predicate_shortfall now resolves to the foundation
proximadb_search_types::PredicateShortfall (step 1). A re-export shim in legacy.rs
keeps every existing caller compiling with zero import-path changes.

Makes the upcoming RecordSearchPort contract self-contained.
