+++ crates/platform/proximadb-runtime/src/lib.rs
@@ -21,6 +21,7 @@ pub mod record_ops_port;
 pub mod record_route_port;
 pub mod resources;
 pub mod rich_record;
+pub mod rich_search;
 pub mod security_port;
 pub mod service_ports;
 pub mod streaming_port;
@@ -46,6 +47,9 @@ pub use record_ops_port::RecordOpsPort;
 pub use record_route_port::{PaxColumnDesc, PaxScanInputs, RecordRoutePort};
 pub use resources::{MemoryBudget, ResourceManager};
 pub use rich_record::{RichRecordBatchRequest, RichRecordDeleteBatchRequest, RichRecordGetRequest};
+pub use rich_search::{
+    RichFilterCondition, RichFilterOperator, RichSearchRequest, RichSearchResponse, RichSearchResult,
+};
 pub use security_port::{PortAuthCredential, PortUserContext, SecurityPort};
 pub use service_ports::{CollectionPort, QueryAdapterPort, VectorOpsPort};
 pub use streaming_port::StreamingPort;
